locking barcodes

+++ application/sql/schema.sql
@@ -813,16 +813,31 @@ end
 $$ language 'plpgsql';
 
 
+create function sites_parse_type()
+  returns trigger as
+$$
+  declare x_site text[];
+begin
+  if new.name is not null then
+    select regexp_matches(new.name::text, E'^([A-Z]{3})([\\s_-]*[A-Z0-9]{1,10})?$') into x_site;
+    new.type = x_site[1];
+  end if;
+
+  return new;
+end
+$$ language 'plpgsql';
+
+
 create function barcodes_hops()
   returns trigger as
 $$
 begin
-  if tg_op = 'INSERT' then
+  if (tg_op = 'INSERT') then
     perform rebuild_barcode_hops(new.id);
-  elseif tg_op = 'UPDATE' then
+  elseif (tg_op = 'UPDATE') and (old.parent_id <> new.parent_id) then
     delete from barcode_hops_cached where barcode_id = old.id;
     perform rebuild_barcode_hops(new.id);
-  elseif tg_op = 'DELETE' then
+  elseif (tg_op = 'DELETE') then
     delete from barcode_hops_cached where barcode_id = old.id;
   end if;
 
@@ -831,7 +846,7 @@ end
 $$ language 'plpgsql';
 
 
-create or replace function rebuild_barcode_hops(x_barcode_id d_id)
+create function rebuild_barcode_hops(x_barcode_id d_id)
   returns void as
 $$
   declare x_id d_id;
@@ -863,41 +878,60 @@ end
 $$ language 'plpgsql';
 
 
--- create function barcodes_locks()
---   returns trigger as
--- $$
--- begin
---   if tg_op = 'INSERT' then
---     if new.invoice_id is not null then
---       perform rebuild_barcode_locks(new.barcode_id, true);
---     end if;
---   elseif tg_op = 'UPDATE' then
---     if (old.invoice_id is null) and (new.invoice_id is not null) then
---       perform rebuild_barcode_locks(new.barcode_id, true);
---     elseif (old.invoice_id is not null) and (new.invoice_id is null) then
---       perform rebuild_barcode_locks(new.barcode_id, false);
---     end if;
---   end if;
---
---   return null;
--- end
--- $$ language 'plpgsql';
+create function check_barcode_locks()
+  returns trigger as
+$$
+  declare x_is_locked d_bool;
+begin
+  select is_locked from barcodes where id = old.barcode_id into x_is_locked;
+
+  if x_is_locked = true then
+    raise exception 'Sorry, cannot modify data due to barcode locks.';
+    return null;
+
+  elseif (tg_op = 'INSERT') or (tg_op = 'UPDATE') then
+    return new;
+
+  elseif (tg_op = 'DELETE') then
+    return old;
+
+  end if;
+
+end
+$$ language 'plpgsql';
 
 
--- create function rebuild_barcode_locks(x_barcode_id d_id, x_locked d_bool)
---   returns void as
--- $$
---   declare x_id d_id;
--- begin
---   for x_id in select barcode_id from barcode_hops_cached where parent_id = x_barcode_id loop
---     update barcodes set is_locked = x_locked where id = x_id;
---   end loop;
---
---   for x_id in select parent_id from barcode_hops_cached where barcode_id = x_barcode_id loop
---     update barcodes set is_locked = x_locked where id = x_id;
---   end loop;
--- end
--- $$ language 'plpgsql';
+create function barcodes_locks()
+  returns trigger as
+$$
+begin
+  if (old.is_locked <> new.is_locked) and (new.is_locked = true) then
+    perform rebuild_barcode_locks(new.id, true);
+  end if;
+
+  return null;
+end
+$$ language 'plpgsql';
+
+
+create function rebuild_barcode_locks(x_barcode_id d_id, x_lock d_bool)
+  returns void as
+$$
+  declare x_id d_id;
+  declare x_is_locked d_bool;
+begin
+  if (x_barcode_id is null) and (x_lock is null) then
+    for x_id in select id from barcodes loop
+      perform rebuild_barcode_locks(x_id, true);
+    end loop;
+
+  else
+    for x_id in select barcode_id from barcode_hops_cached where parent_id = x_barcode_id loop
+      update barcodes set is_locked = x_lock where id = x_id;
+    end loop;
+  end if;
+end
+$$ language 'plpgsql';
 
 
 create function ssf_data_update_barcodes()
@@ -1034,17 +1068,40 @@ end
 $$ language 'plpgsql';
 
 
-create function sites_parse_type()
+create function invoice_data_update_barcodes()
   returns trigger as
 $$
-  declare x_site text[];
+  declare x_barcode_id d_id;
+  declare x_record record;
+  declare x_form_type d_form_type;
+  declare x_form_data_id d_id;
 begin
-  if new.name is not null then
-    select regexp_matches(new.name::text, E'^([A-Z]{3})([\\s_-]*[A-Z0-9]{1,10})?$') into x_site;
-    new.type = x_site[1];
+  if (tg_op = 'DELETE') then
+    select old.form_type into x_form_type;
+    select old.form_data_id into x_form_data_id;
+  else
+    select new.form_type into x_form_type;
+    select new.form_data_id into x_form_data_id;
   end if;
 
-  return new;
+  case x_form_type
+    when 'SSF'   then select barcode_id from ssf_data where id = x_form_data_id into x_barcode_id;
+    when 'TDF'   then select barcode_id from tdf_data where id = x_form_data_id into x_barcode_id;
+    when 'LDF'   then select barcode_id from ldf_data where id = x_form_data_id into x_barcode_id;
+    when 'MIF'   then select barcode_id from mif_data where id = x_form_data_id into x_barcode_id;
+    when 'MOF'   then select barcode_id from mof_data where id = x_form_data_id into x_barcode_id;
+    when 'SPECS' then select barcode_id from specs_data where id = x_form_data_id into x_barcode_id;
+    when 'EPR'   then select barcode_id from epr_data where id = x_form_data_id into x_barcode_id;
+    else return null;
+  end case;
+
+  if (tg_op = 'DELETE') then
+    update barcodes set is_locked = false where id = x_barcode_id;
+  else
+    update barcodes set is_locked = true where id = x_barcode_id;
+  end if;
+
+  return null;
 end
 $$ language 'plpgsql';
 
@@ -1064,6 +1121,11 @@ $$ language 'plpgsql';
 
 
 -- triggers
+
+create trigger t_sites_parse_type
+  before insert or update on sites
+  for each row
+  execute procedure sites_parse_type();
 
 create trigger t_barcodes_hops
   after insert or delete or update on barcodes
@@ -1095,15 +1157,50 @@ create trigger t_specs_data_update_barcodes
   for each row
   execute procedure specs_data_update_barcodes();
 
--- create trigger t_barcodes_locks
---   after insert or update on ldf_data
---   for each row
---   execute procedure barcodes_locks();
-
-create trigger t_sites_parse_type
-  before insert or update on sites
+create trigger t_barcodes_locks
+  after insert or update on barcodes
   for each row
-  execute procedure sites_parse_type();
+  execute procedure barcodes_locks();
+
+create trigger t_check_barcode_locks
+  before insert or update or delete on ssf_data
+  for each row
+  execute procedure check_barcode_locks();
+
+create trigger t_check_barcode_locks
+  before insert or update or delete on tdf_data
+  for each row
+  execute procedure check_barcode_locks();
+
+create trigger t_check_barcode_locks
+  before insert or update or delete on ldf_data
+  for each row
+  execute procedure check_barcode_locks();
+
+create trigger t_check_barcode_locks
+  before insert or update or delete on mof_data
+  for each row
+  execute procedure check_barcode_locks();
+
+create trigger t_check_barcode_locks
+  before insert or update or delete on mif_data
+  for each row
+  execute procedure check_barcode_locks();
+
+create trigger t_check_barcode_locks
+  before insert or update or delete on specs_data
+  for each row
+  execute procedure check_barcode_locks();
+
+create trigger t_check_barcode_locks
+  before insert or update or delete on epr_data
+  for each row
+  execute procedure check_barcode_locks();
+
+create trigger t_invoice_data_update_barcodes
+  after insert or update or delete on invoice_data
+  for each row
+  execute procedure invoice_data_update_barcodes();
 
 -- create trigger t_csv_integrity
 --   before update on csv
